articleId and teamId now string and used in safeArgs instead of SharedViewModel

+++ app/src/main/java/com/tryden/simplenfl/domain/mappers/teamslist/UiTeamMapper.kt
@@ -6,7 +6,7 @@ import com.tryden.simplenfl.domain.models.teamslist.UiTeam
 class UiTeamMapper {
     fun buildFrom(team: Team): UiTeam {
         return UiTeam(
-            id = team.id.toInt(),
+            id = team.id,
             imageUrl = team.logos[0].href,
             name = team.shortDisplayName
         )
